HSEARCH-4144 Refactor ORM entity loading classes and move them to a dedicated package
So that we can start using them for something else than just searching.
