+++ mapper/orm/src/main/java/org/hibernate/search/mapper/orm/loading/impl/EntityLoadingCacheLookupStrategyImplementor.java
@@ -4,11 +4,11 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.mapper.orm.search.loading.impl;
+package org.hibernate.search.mapper.orm.loading.impl;
 
 import org.hibernate.engine.spi.EntityKey;
 
-public interface EntityLoadingCacheLookupStrategyImplementor {
+interface EntityLoadingCacheLookupStrategyImplementor {
 
 	/**
 	 * @param entityKey The key of an entity.
