+++ mapper/orm/src/main/java/org/hibernate/search/mapper/orm/loading/impl/EntityLoadingStrategy.java
@@ -1,0 +1,13 @@
+/*
+ * Hibernate Search, full-text search for your domain model
+ *
+ * License: GNU Lesser General Public License (LGPL), version 2.1 or later
+ * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
+ */
+package org.hibernate.search.mapper.orm.loading.impl;
+
+import org.hibernate.search.mapper.orm.search.loading.impl.SearchEntityLoadingStrategy;
+
+public interface EntityLoadingStrategy extends SearchEntityLoadingStrategy {
+
+}
