+++ mapper/orm/src/main/java/org/hibernate/search/mapper/orm/loading/impl/HibernateOrmEntityIdEntityLoader.java
@@ -4,7 +4,7 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.mapper.orm.search.loading.impl;
+package org.hibernate.search.mapper.orm.loading.impl;
 
 import java.io.Serializable;
 import java.util.ArrayList;
@@ -12,32 +12,25 @@ import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 
-import org.hibernate.AssertionFailure;
 import org.hibernate.QueryTimeoutException;
 import org.hibernate.engine.spi.EntityKey;
-import org.hibernate.engine.spi.SessionFactoryImplementor;
 import org.hibernate.engine.spi.SessionImplementor;
 import org.hibernate.exception.LockTimeoutException;
 import org.hibernate.jpa.QueryHints;
-import org.hibernate.metamodel.spi.MetamodelImplementor;
 import org.hibernate.persister.entity.EntityPersister;
 import org.hibernate.query.Query;
 import org.hibernate.search.mapper.orm.common.EntityReference;
-import org.hibernate.search.mapper.orm.common.impl.HibernateOrmUtils;
-import org.hibernate.search.mapper.orm.search.loading.EntityLoadingCacheLookupStrategy;
 import org.hibernate.search.engine.common.timing.spi.Deadline;
+import org.hibernate.search.mapper.orm.search.loading.impl.HibernateOrmComposableSearchEntityLoader;
+import org.hibernate.search.mapper.orm.search.loading.impl.EntityGraphHint;
+import org.hibernate.search.mapper.orm.search.loading.impl.MutableEntityLoadingOptions;
 
 /**
  * An entity loader for indexed entities whose document ID is the entity ID.
  *
  * @param <E> The type of loaded entities.
  */
-public class HibernateOrmEntityIdEntityLoader<E> implements HibernateOrmComposableEntityLoader<E> {
-
-	public static EntityLoaderFactory factory(SessionFactoryImplementor sessionFactory,
-			EntityPersister entityPersister) {
-		return new Factory( HibernateOrmUtils.toRootEntityType( sessionFactory, entityPersister ) );
-	}
+class HibernateOrmEntityIdEntityLoader<E> implements HibernateOrmComposableSearchEntityLoader<E> {
 
 	private static final String IDS_PARAMETER_NAME = "ids";
 
@@ -47,7 +40,7 @@ public class HibernateOrmEntityIdEntityLoader<E> implements HibernateOrmComposab
 	private final EntityLoadingCacheLookupStrategyImplementor cacheLookupStrategyImplementor;
 	private final MutableEntityLoadingOptions loadingOptions;
 
-	private HibernateOrmEntityIdEntityLoader(
+	HibernateOrmEntityIdEntityLoader(
 			EntityPersister entityPersister,
 			SessionImplementor session,
 			PersistenceContextLookupStrategy persistenceContextLookup,
@@ -78,7 +71,7 @@ public class HibernateOrmEntityIdEntityLoader<E> implements HibernateOrmComposab
 			}
 		}
 		else {
-			return HibernateOrmComposableEntityLoader.super.loadBlocking( references, deadline );
+			return HibernateOrmComposableSearchEntityLoader.super.loadBlocking( references, deadline );
 		}
 	}
 
@@ -171,7 +164,7 @@ public class HibernateOrmEntityIdEntityLoader<E> implements HibernateOrmComposab
 	}
 
 	private Query<?> createQuery(int fetchSize, Long timeout) {
-		Query<?> query = HibernateOrmUtils.createQueryForLoadByUniqueProperty(
+		Query<?> query = HibernateOrmQueryUtils.createQueryForLoadByUniqueProperty(
 				session, entityPersister, entityPersister.getIdentifierPropertyName(), IDS_PARAMETER_NAME
 		);
 
@@ -234,135 +227,4 @@ public class HibernateOrmEntityIdEntityLoader<E> implements HibernateOrmComposab
 		return reference.type().isInstance( loadedEntity );
 	}
 
-	private static class Factory implements EntityLoaderFactory {
-
-		private final EntityPersister rootEntityPersister;
-
-		private Factory(EntityPersister rootEntityPersister) {
-			this.rootEntityPersister = rootEntityPersister;
-		}
-
-		@Override
-		public boolean equals(Object obj) {
-			if ( obj == null || !( getClass().equals( obj.getClass() ) ) ) {
-				return false;
-			}
-			Factory other = (Factory) obj;
-			// If the root entity type is different,
-			// the factories work in separate ID spaces and should be used separately.
-			return rootEntityPersister.equals( other.rootEntityPersister );
-		}
-
-		@Override
-		public int hashCode() {
-			return rootEntityPersister.hashCode();
-		}
-
-		@Override
-		public <E> HibernateOrmComposableEntityLoader<E> create(
-				HibernateOrmLoadingIndexedTypeContext targetEntityTypeContext,
-				SessionImplementor session,
-				EntityLoadingCacheLookupStrategy cacheLookupStrategy, MutableEntityLoadingOptions loadingOptions) {
-			/*
-			 * This cast is safe: the loader will only return instances of E.
-			 * See hasExpectedType() and its callers for more information,
-			 * in particular runtime checks handling edge cases.
-			 */
-			@SuppressWarnings("unchecked")
-			HibernateOrmComposableEntityLoader<E> result = (HibernateOrmComposableEntityLoader<E>) doCreate(
-					targetEntityTypeContext.entityPersister(), session, cacheLookupStrategy, loadingOptions
-			);
-			return result;
-		}
-
-		@Override
-		public <E> HibernateOrmComposableEntityLoader<? extends E> create(
-				List<HibernateOrmLoadingIndexedTypeContext> targetEntityTypeContexts,
-				SessionImplementor session, EntityLoadingCacheLookupStrategy cacheLookupStrategy,
-				MutableEntityLoadingOptions loadingOptions) {
-			EntityPersister commonSuperType = toMostSpecificCommonEntitySuperType( session, targetEntityTypeContexts );
-
-			/*
-			 * Theoretically, this cast is unsafe,
-			 * since the loader could return entities of any type T extending "commonSuperClass",
-			 * which is either E (good: T = E)
-			 * or a common supertype of some child types of E
-			 * (not good: T might be an interface that E doesn't implement but its children do).
-			 *
-			 * However, we perform some runtime checks that make this cast safe.
-			 *
-			 * See hasExpectedType() and its callers for more information.
-			 */
-			@SuppressWarnings("unchecked")
-			HibernateOrmComposableEntityLoader<E> result = (HibernateOrmComposableEntityLoader<E>) doCreate(
-					commonSuperType, session, cacheLookupStrategy, loadingOptions
-			);
-
-			return result;
-		}
-
-		private HibernateOrmComposableEntityLoader<?> doCreate(EntityPersister entityPersister,
-				SessionImplementor session,
-				EntityLoadingCacheLookupStrategy cacheLookupStrategy, MutableEntityLoadingOptions loadingOptions) {
-			if ( !rootEntityPersister.getMappedClass().isAssignableFrom( entityPersister.getMappedClass() ) ) {
-				throw new AssertionFailure(
-						"Some types among the targeted entity types are not subclasses of the expected root entity type."
-								+ " There is a bug in Hibernate Search, please report it."
-								+ " Expected root entity name: " + rootEntityPersister.getEntityName()
-								+ " Targeted entity name: " + entityPersister.getEntityName()
-				);
-			}
-
-			PersistenceContextLookupStrategy persistenceContextLookup =
-					PersistenceContextLookupStrategy.create( session );
-			EntityLoadingCacheLookupStrategyImplementor cacheLookupStrategyImplementor;
-
-			/*
-			 * Ideally, in order to comply with the cache lookup strategy,
-			 * we would use multiAccess setters such as
-			 * with(CacheMode) and enableSessionCheck(boolean),
-			 * and let Hibernate ORM do it for us.
-			 *
-			 * However, with(CacheMode) has a side-effect: it can also affect how entities are put into the cache.
-			 * Since the cache lookup strategy has nothing to do with that,
-			 * we go the safer route and wrap the loader with other loaders that
-			 * will perform PC and 2LC checking prior to using the multiAccess.
-			 */
-			switch ( cacheLookupStrategy ) {
-				case SKIP:
-					cacheLookupStrategyImplementor = null;
-					break;
-				case PERSISTENCE_CONTEXT:
-					cacheLookupStrategyImplementor = persistenceContextLookup;
-					break;
-				case PERSISTENCE_CONTEXT_THEN_SECOND_LEVEL_CACHE:
-					cacheLookupStrategyImplementor =
-							PersistenceContextThenSecondLevelCacheLookupStrategy.create( entityPersister, session );
-					break;
-				default:
-					throw new AssertionFailure( "Unexpected cache lookup strategy: " + cacheLookupStrategy );
-			}
-
-			return new HibernateOrmEntityIdEntityLoader<>(
-					entityPersister, session, persistenceContextLookup, cacheLookupStrategyImplementor, loadingOptions
-			);
-		}
-
-		private static EntityPersister toMostSpecificCommonEntitySuperType(SessionImplementor session,
-				Iterable<? extends HibernateOrmLoadingIndexedTypeContext> targetEntityTypeContexts) {
-			MetamodelImplementor metamodel = session.getSessionFactory().getMetamodel();
-			EntityPersister result = null;
-			for ( HibernateOrmLoadingIndexedTypeContext targetTypeContext : targetEntityTypeContexts ) {
-				EntityPersister type = targetTypeContext.entityPersister();
-				if ( result == null ) {
-					result = type;
-				}
-				else {
-					result = HibernateOrmUtils.toMostSpecificCommonEntitySuperType( metamodel, result, type );
-				}
-			}
-			return result;
-		}
-
-	}
 }
